refactor: move useFrameworkReady hook into src
