+++ hooks/useFrameworkReady.ts
@@ -1,13 +1,1 @@
-import { useEffect } from 'react';
-
-declare global {
-  interface Window {
-    frameworkReady?: () => void;
-  }
-}
-
-export function useFrameworkReady() {
-  useEffect(() => {
-    window.frameworkReady?.()
-  })
-}
+export { useFrameworkReady } from '@/hooks/useFrameworkReady';
